Tools - reorganize packages
