+++ .phpqa.yml
@@ -69,14 +69,14 @@ report:
 
 # you can add new tools, but existing tools cannot be removed
 tool:
-    phpmetrics: [Edge\QA\Tool\PhpMetrics, Edge\QA\Tool\PhpMetricsV2]
-    phploc: Edge\QA\Tool\Phploc
-    phpcs: [Edge\QA\Tool\Phpcs, Edge\QA\Tool\PhpcsV3]
-    php-cs-fixer: Edge\QA\Tool\PhpCsFixer
-    phpmd: Edge\QA\Tool\Phpmd
-    pdepend: Edge\QA\Tool\Pdepend
-    phpcpd: Edge\QA\Tool\Phpcpd
-    parallel-lint: Edge\QA\Tool\ParallelLint
-    phpstan: Edge\QA\Tool\Phpstan
-    phpunit: Edge\QA\Tool\Phpunit
-    psalm: Edge\QA\Tool\Psalm
+    phpmetrics: [Edge\QA\Tools\Analyzer\PhpMetrics, Edge\QA\Tools\Analyzer\PhpMetricsV2]
+    phploc: Edge\QA\Tools\Analyzer\Phploc
+    phpcs: [Edge\QA\Tools\Analyzer\Phpcs, Edge\QA\Tools\Analyzer\PhpcsV3]
+    php-cs-fixer: Edge\QA\Tools\Analyzer\PhpCsFixer
+    phpmd: Edge\QA\Tools\Analyzer\Phpmd
+    pdepend: Edge\QA\Tools\Analyzer\Pdepend
+    phpcpd: Edge\QA\Tools\Analyzer\Phpcpd
+    parallel-lint: Edge\QA\Tools\Analyzer\ParallelLint
+    phpstan: Edge\QA\Tools\Analyzer\Phpstan
+    phpunit: Edge\QA\Tools\Analyzer\Phpunit
+    psalm: Edge\QA\Tools\Analyzer\Psalm
