+++ src/Tools/Analyzer/Pdepend.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Edge\QA\Tool;
+namespace Edge\QA\Tools\Analyzer;
 
-class Pdepend extends Tool
+class Pdepend extends \Edge\QA\Tools\Tool
 {
     public static $SETTINGS = array(
         'optionSeparator' => '=',
