+++ src/Tools/Analyzer/PhpMetricsV2.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Edge\QA\Tool;
+namespace Edge\QA\Tools\Analyzer;
 
-class PhpMetricsV2 extends Tool
+class PhpMetricsV2 extends \Edge\QA\Tools\Tool
 {
     public static $SETTINGS = array(
         'optionSeparator' => '=',
