+++ src/Tools/Analyzer/Phpcpd.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Edge\QA\Tool;
+namespace Edge\QA\Tools\Analyzer;
 
-class Phpcpd extends Tool
+class Phpcpd extends \Edge\QA\Tools\Tool
 {
     public static $SETTINGS = array(
         'optionSeparator' => ' ',
