+++ src/Tools/Analyzer/Phpmd.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Edge\QA\Tool;
+namespace Edge\QA\Tools\Analyzer;
 
-class Phpmd extends Tool
+class Phpmd extends \Edge\QA\Tools\Tool
 {
     public static $SETTINGS = array(
         'optionSeparator' => ' ',
