+++ src/Tools/Analyzer/Psalm.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Edge\QA\Tool;
+namespace Edge\QA\Tools\Analyzer;
 
-class Psalm extends Tool
+class Psalm extends \Edge\QA\Tools\Tool
 {
     public static $SETTINGS = array(
         'optionSeparator' => '=',
@@ -10,13 +10,12 @@ class Psalm extends Tool
         'errorsXPath' => '//item/severity[text()=\'error\']',
         'composer' => 'vimeo/psalm',
         'internalClass' => 'Psalm\Checker\ProjectChecker',
-        'handler' => 'Edge\QA\Tool\Psalm',
     );
 
     public function __invoke()
     {
         if (!$this->config->value('psalm.config')) {
-            $twig = new \Twig_Environment(new \Twig_Loader_Filesystem(__DIR__.'/../../app/'));
+            $twig = new \Twig_Environment(new \Twig_Loader_Filesystem(__DIR__ . '/../../../app/'));
             $psalmXml = $twig->render(
                 'psalm.xml.twig',
                 array(
